add render_mime_structure function and test
